[MNT] address some warnings and deprecation messages from dependencies (#4486)
This PR addresses some warnings deprecation messages from dependencies:

* default `dtype` set explicitly in `pandas` constructors to `float64`
* adding `np.abs` to prevent negative values in square roots

+++ sktime/forecasting/squaring_residuals.py
@@ -350,7 +350,7 @@ class SquaringResiduals(BaseForecaster):
         fh_abs = fh.to_absolute(self.cutoff)
         fh_rel = fh.to_relative(self.cutoff)
         fh_rel_index = fh_rel.to_pandas()
-        pred_var = pd.Series(index=fh_rel_index)
+        pred_var = pd.Series(index=fh_rel_index, dtype="float64")
         for el in fh_rel:
             pred_var.at[el] = self._res_forecasters[el].predict(fh=el)[0]
         if self.strategy == "square":
